refactor: extract MessageBuilder, ReviewHandler, ApprovalHandler from AiAgentService
- AiAgent::MessageBuilder: builds message array (creative context, chat history, trigger)
- AiAgent::ReviewHandler: review eligibility, in-place updates, completion reactions
- AiAgent::ApprovalHandler: tool approval flow (placeholder cleanup, task update, approval comment)
- AiAgentService reduced from 458 to 251 lines
- All 8 existing tests pass, 4 new ReviewHandler tests added
